requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4
Keras-Preprocessing==1.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
opt-einsum==3.3.0
pyasn1-modules==0.3.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
tensorboard==2.10.1
tensorflow==2.13.0
tensorflow-intel==2.13.0
typing_extensions==4.7.1
tzdata==2023.3
Werkzeug==2.3.6
